# docs/design/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SyncFlow",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/info": {
      "get": {
        "summary": "Project Info",
        "operationId": "project_info_api_v1_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_dict_str__str__"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "INVALID_REQUEST": {
                    "value": {
                      "error": {
                        "code": "INVALID_REQUEST",
                        "message": "请求参数不合法",
                        "details": []
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "error": {
                        "code": "INTERNAL_ERROR",
                        "message": "服务内部错误，请稍后重试",
                        "details": []
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs": {
      "post": {
        "summary": "创建任务",
        "description": "校验并保存 CSV、提交 PENDING 任务并投递 Redis，不等待 Worker。默认限制 10 MiB，扩展名大小写不敏感。投递失败记录任务错误并返回 500。",
        "operationId": "createJob",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_createJob"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_JobCreated_"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST, INVALID_FILE_EXTENSION, FILE_TOO_LARGE",
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_REQUEST": {
                    "value": {
                      "error": {
                        "code": "INVALID_REQUEST",
                        "message": "请求参数不合法",
                        "details": []
                      }
                    }
                  },
                  "INVALID_FILE_EXTENSION": {
                    "value": {
                      "error": {
                        "code": "INVALID_FILE_EXTENSION",
                        "message": "仅支持 CSV 文件",
                        "details": []
                      }
                    }
                  },
                  "FILE_TOO_LARGE": {
                    "value": {
                      "error": {
                        "code": "FILE_TOO_LARGE",
                        "message": "文件超过大小限制",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "error": {
                        "code": "INTERNAL_ERROR",
                        "message": "服务内部错误，请稍后重试",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "查询任务列表",
        "description": "支持 page（默认 1）、page_size（默认 20，最大 100）与状态筛选；按 created_at DESC, id DESC 排序。返回公开任务字段及 page/page_size/total。",
        "operationId": "listJobs",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "PENDING",
                    "RUNNING",
                    "RETRYING",
                    "SUCCESS",
                    "PARTIAL_SUCCESS",
                    "FAILED",
                    "CANCELING",
                    "CANCELED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_REQUEST": {
                    "value": {
                      "error": {
                        "code": "INVALID_REQUEST",
                        "message": "请求参数不合法",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "error": {
                        "code": "INTERNAL_ERROR",
                        "message": "服务内部错误，请稍后重试",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "summary": "查询任务详情",
        "description": "返回任务公开字段、统计与 UTC 时间；不包含内部存储路径。不存在返回 JOB_NOT_FOUND。",
        "operationId": "getJob",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 36,
              "description": "UUID 或 ULID 任务标识",
              "title": "Job Id"
            },
            "description": "UUID 或 ULID 任务标识"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_JobDetail_"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_REQUEST": {
                    "value": {
                      "error": {
                        "code": "INVALID_REQUEST",
                        "message": "请求参数不合法",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "error": {
                        "code": "INTERNAL_ERROR",
                        "message": "服务内部错误，请稍后重试",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND",
            "content": {
              "application/json": {
                "examples": {
                  "JOB_NOT_FOUND": {
                    "value": {
                      "error": {
                        "code": "JOB_NOT_FOUND",
                        "message": "任务不存在",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "summary": "取消任务",
        "description": "等待执行或重试的任务直接取消；执行中的任务先进入 CANCELING，Worker 停止执行后进入 CANCELED。已请求取消或终态返回 409，保留已提交数据。",
        "operationId": "cancelJob",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 36,
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_JobDetail_"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_REQUEST": {
                    "value": {
                      "error": {
                        "code": "INVALID_REQUEST",
                        "message": "请求参数不合法",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "error": {
                        "code": "INTERNAL_ERROR",
                        "message": "服务内部错误，请稍后重试",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND",
            "content": {
              "application/json": {
                "examples": {
                  "JOB_NOT_FOUND": {
                    "value": {
                      "error": {
                        "code": "JOB_NOT_FOUND",
                        "message": "任务不存在",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "INVALID_JOB_TRANSITION",
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_JOB_TRANSITION": {
                    "value": {
                      "error": {
                        "code": "INVALID_JOB_TRANSITION",
                        "message": "任务状态不允许此操作",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/errors": {
      "get": {
        "summary": "查询任务错误明细",
        "description": "支持 page（默认 1）与 page_size（默认 20，最大 100）。按 row_number ASC, id ASC 排序，文件级空行号排在前面；返回原始 JSON 和 UTC 时间。任务不存在返回 JOB_NOT_FOUND，任务存在但无错误返回空列表，运行中也可查询。",
        "operationId": "listJobErrors",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 36,
              "description": "UUID 或 ULID 任务标识",
              "title": "Job Id"
            },
            "description": "UUID 或 ULID 任务标识"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "错误明细与分页信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobErrorListResponse"
                },
                "examples": {
                  "row_error": {
                    "summary": "行级错误",
                    "value": {
                      "data": [
                        {
                          "job_id": "example-job",
                          "row_number": 2,
                          "field_name": "amount",
                          "error_code": "AMOUNT_INVALID",
                          "error_message": "amount 不能为负数",
                          "raw_row": [
                            "S001",
                            "商品A",
                            "-1",
                            "2026-01-01"
                          ],
                          "created_at": "2026-01-01T10:00:00Z"
                        }
                      ],
                      "meta": {
                        "page": 1,
                        "page_size": 20,
                        "total": 1
                      }
                    }
                  },
                  "file_error": {
                    "summary": "文件级错误",
                    "value": {
                      "data": [
                        {
                          "job_id": "example-job",
                          "row_number": null,
                          "field_name": null,
                          "error_code": "FILE_EMPTY",
                          "error_message": "文件为空或只有表头",
                          "raw_row": null,
                          "created_at": "2026-01-01T10:00:00Z"
                        }
                      ],
                      "meta": {
                        "page": 1,
                        "page_size": 20,
                        "total": 1
                      }
                    }
                  },
                  "empty": {
                    "summary": "任务存在但无错误",
                    "value": {
                      "data": [],
                      "meta": {
                        "page": 1,
                        "page_size": 20,
                        "total": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_REQUEST": {
                    "value": {
                      "error": {
                        "code": "INVALID_REQUEST",
                        "message": "请求参数不合法",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "error": {
                        "code": "INTERNAL_ERROR",
                        "message": "服务内部错误，请稍后重试",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "JOB_NOT_FOUND",
            "content": {
              "application/json": {
                "examples": {
                  "JOB_NOT_FOUND": {
                    "value": {
                      "error": {
                        "code": "JOB_NOT_FOUND",
                        "message": "任务不存在",
                        "details": []
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_createJob": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "CSV 文件，默认不超过 10 MiB"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "不传时由系统生成任务名称"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_createJob"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorBody"
      },
      "ErrorDetail": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "JobCreated": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 36,
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 128,
            "title": "Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "RETRYING",
              "SUCCESS",
              "PARTIAL_SUCCESS",
              "FAILED",
              "CANCELING",
              "CANCELED"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "created_at"
        ],
        "title": "JobCreated"
      },
      "JobDetail": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 36,
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 128,
            "title": "Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "RETRYING",
              "SUCCESS",
              "PARTIAL_SUCCESS",
              "FAILED",
              "CANCELING",
              "CANCELED"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "source_file_name": {
            "type": "string",
            "title": "Source File Name"
          },
          "total_records": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Records"
          },
          "success_records": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Success Records"
          },
          "failed_records": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Failed Records"
          },
          "retry_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Retry Count"
          },
          "last_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Code"
          },
          "last_error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Message"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "created_at",
          "source_file_name",
          "total_records",
          "success_records",
          "failed_records",
          "retry_count",
          "last_error_code",
          "last_error_message",
          "started_at",
          "finished_at"
        ],
        "title": "JobDetail"
      },
...
      "JobErrorListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/JobError"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "JobErrorListResponse"
      },
      "JobListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/JobDetail"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "JobListResponse"
      },
...
      "PageMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total"
        ],
        "title": "PageMeta"
      },
...
      "SuccessResponse_JobCreated_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobCreated"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageMeta"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[JobCreated]"
      },
      "SuccessResponse_JobDetail_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobDetail"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageMeta"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[JobDetail]"
      },
      "SuccessResponse_dict_str__str__": {
        "properties": {
          "data": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Data"
          },
          "meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageMeta"
              },
              {
                "additionalProperties": true,
                "type": "object"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[dict[str, str]]"
      },
...
    }
  }
}